Modularize API workflow layer
