+++ crates/pcl/cli/tests/auth_output.rs
@@ -1148,224 +1148,6 @@ fn completions_can_run_with_invalid_config_without_overwriting_file() {
         original_config
     );
 }
-
-#[test]
-fn api_dry_run_project_create_does_not_hit_network() {
-    let temp_dir = tempfile::tempdir().expect("create temp config dir");
-    write_valid_auth_config(temp_dir.path());
-
-    let output = Command::new(env!("CARGO_BIN_EXE_pcl"))
-        .args([
-            "--config-dir",
-            temp_dir.path().to_str().expect("utf-8 temp path"),
-            "--json",
-            "api",
-            "--api-url",
-            "http://127.0.0.1:9",
-            "--dry-run",
-            "projects",
-            "--create",
-            "--project-name",
-            "demo",
-            "--chain-id",
-            "1",
-        ])
-        .output()
-        .expect("run pcl api projects dry-run");
-
-    assert!(
-        output.status.success(),
-        "dry-run attempted network/auth path: {}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-    let stdout = String::from_utf8(output.stdout).expect("utf-8 stdout");
-    let envelope: serde_json::Value = serde_json::from_str(&stdout).expect("json envelope");
-    assert_eq!(envelope["status"], "ok");
-    assert_eq!(envelope["data"]["dry_run"], true);
-    assert_eq!(envelope["data"]["request"]["method"], "POST");
-    assert_eq!(envelope["data"]["request"]["path"], "/projects");
-}
-
-#[test]
-fn api_dry_run_assertion_registered_does_not_hit_network() {
-    let temp_dir = tempfile::tempdir().expect("create temp config dir");
-    write_valid_auth_config(temp_dir.path());
-
-    let output = Command::new(env!("CARGO_BIN_EXE_pcl"))
-        .args([
-            "--config-dir",
-            temp_dir.path().to_str().expect("utf-8 temp path"),
-            "--json",
-            "api",
-            "--api-url",
-            "http://127.0.0.1:9",
-            "--dry-run",
-            "assertions",
-            "--project-id",
-            "project-1",
-            "--registered",
-        ])
-        .output()
-        .expect("run pcl api assertions dry-run");
-
-    assert!(
-        output.status.success(),
-        "dry-run attempted network/auth path: {}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-    let stdout = String::from_utf8(output.stdout).expect("utf-8 stdout");
-    let envelope: serde_json::Value = serde_json::from_str(&stdout).expect("json envelope");
-    assert_eq!(envelope["status"], "ok");
-    assert_eq!(envelope["data"]["dry_run"], true);
-    assert_eq!(envelope["data"]["request"]["method"], "GET");
-    assert_eq!(
-        envelope["data"]["request"]["path"],
-        "/projects/project-1/registered-assertions"
-    );
-}
-
-#[test]
-fn api_dry_run_auth_metadata_keeps_required_separate_from_attachment() {
-    let temp_dir = tempfile::tempdir().expect("create temp config dir");
-    write_valid_auth_config(temp_dir.path());
-
-    let output = Command::new(env!("CARGO_BIN_EXE_pcl"))
-        .args([
-            "--config-dir",
-            temp_dir.path().to_str().expect("utf-8 temp path"),
-            "--json",
-            "api",
-            "--allow-unauthenticated",
-            "--dry-run",
-            "projects",
-            "--home",
-        ])
-        .output()
-        .expect("run pcl api dry-run");
-
-    assert!(
-        output.status.success(),
-        "command failed: {}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-    let stdout = String::from_utf8(output.stdout).expect("utf-8 stdout");
-    let envelope: serde_json::Value = serde_json::from_str(&stdout).expect("json envelope");
-    assert_eq!(envelope["data"]["request"]["auth"]["required"], true);
-    assert_eq!(
-        envelope["data"]["request"]["auth"]["will_attach_stored_token"],
-        false
-    );
-}
-
-#[test]
-fn top_level_project_workflow_matches_api_alias() {
-    let temp_dir = tempfile::tempdir().expect("create temp config dir");
-    write_valid_auth_config(temp_dir.path());
-
-    let api_output = Command::new(env!("CARGO_BIN_EXE_pcl"))
-        .args([
-            "--config-dir",
-            temp_dir.path().to_str().expect("utf-8 temp path"),
-            "--json",
-            "api",
-            "--api-url",
-            "http://127.0.0.1:9",
-            "--dry-run",
-            "projects",
-            "--create",
-            "--project-name",
-            "demo",
-            "--chain-id",
-            "1",
-        ])
-        .output()
-        .expect("run pcl api projects dry-run");
-
-    let top_level_output = Command::new(env!("CARGO_BIN_EXE_pcl"))
-        .args([
-            "--config-dir",
-            temp_dir.path().to_str().expect("utf-8 temp path"),
-            "--json",
-            "projects",
-            "--api-url",
-            "http://127.0.0.1:9",
-            "--dry-run",
-            "--create",
-            "--project-name",
-            "demo",
-            "--chain-id",
-            "1",
-        ])
-        .output()
-        .expect("run pcl projects dry-run");
-
-    assert!(
-        api_output.status.success(),
-        "api alias failed: {}",
-        String::from_utf8_lossy(&api_output.stderr)
-    );
-    assert!(
-        top_level_output.status.success(),
-        "top-level workflow failed: {}",
-        String::from_utf8_lossy(&top_level_output.stderr)
-    );
-    let api_envelope: serde_json::Value =
-        serde_json::from_slice(&api_output.stdout).expect("api json envelope");
-    let top_level_envelope: serde_json::Value =
-        serde_json::from_slice(&top_level_output.stdout).expect("top-level json envelope");
-    assert_eq!(top_level_envelope["status"], "ok");
-    assert_eq!(top_level_envelope["data"], api_envelope["data"]);
-}
-
-#[test]
-fn top_level_public_incidents_workflow_matches_api_alias() {
-    let temp_dir = tempfile::tempdir().expect("create temp config dir");
-
-    let api_output = Command::new(env!("CARGO_BIN_EXE_pcl"))
-        .args([
-            "--config-dir",
-            temp_dir.path().to_str().expect("utf-8 temp path"),
-            "--json",
-            "api",
-            "--dry-run",
-            "incidents",
-            "--limit",
-            "5",
-        ])
-        .output()
-        .expect("run pcl api incidents dry-run");
-
-    let top_level_output = Command::new(env!("CARGO_BIN_EXE_pcl"))
-        .args([
-            "--config-dir",
-            temp_dir.path().to_str().expect("utf-8 temp path"),
-            "--json",
-            "incidents",
-            "--dry-run",
-            "--limit",
-            "5",
-        ])
-        .output()
-        .expect("run pcl incidents dry-run");
-
-    assert!(
-        api_output.status.success(),
-        "api alias failed: {}",
-        String::from_utf8_lossy(&api_output.stderr)
-    );
-    assert!(
-        top_level_output.status.success(),
-        "top-level workflow failed: {}",
-        String::from_utf8_lossy(&top_level_output.stderr)
-    );
-    let api_envelope: serde_json::Value =
-        serde_json::from_slice(&api_output.stdout).expect("api json envelope");
-    let top_level_envelope: serde_json::Value =
-        serde_json::from_slice(&top_level_output.stdout).expect("top-level json envelope");
-    assert_eq!(top_level_envelope["status"], "ok");
-    assert_eq!(top_level_envelope["data"], api_envelope["data"]);
-}
-
 #[test]
 fn agent_product_surfaces_emit_json_envelopes() {
     let temp_dir = tempfile::tempdir().expect("create temp config dir");
