+++ crates/pcl/core/src/api.rs
@@ -9,51 +9,40 @@
 
 use crate::{
     DEFAULT_PLATFORM_URL,
-    auth::refresh_stored_auth,
     config::CliConfig,
-    error::AuthError,
 };
 use clap::{
     ArgGroup,
     ValueEnum,
 };
-use pcl_common::args::{
-    CliArgs,
-    OutputMode,
-};
-use reqwest::header::{
-    HeaderMap,
-    HeaderName,
-    HeaderValue,
-};
-use serde_json::{
-    Map,
-    Value,
-    json,
-};
+use pcl_common::args::CliArgs;
 use std::{
     cell::Cell,
-    fs,
-    io::Read,
-    path::{
-        Path,
-        PathBuf,
-    },
-    str::FromStr,
+    path::PathBuf,
 };
 
 mod definitions;
+mod envelopes;
+mod error;
+mod input;
 mod manifest;
+mod method;
 mod openapi;
+mod operations;
 mod render;
+mod runner;
+mod runtime_types;
 mod spec;
 mod templates;
+mod transport;
+mod workflow_options;
 mod workflows;
 
 pub use crate::output::{
     ENVELOPE_SCHEMA_VERSION,
     with_envelope_metadata,
 };
+pub use error::ApiCommandError;
 pub use manifest::api_manifest;
 pub use render::{
     envelope_output_string,
@@ -61,10 +50,25 @@ pub use render::{
     toon_string,
 };
 
-use definitions::{
-    WorkflowOutputPolicy,
-    workflow_output_policy,
+use envelopes::{
+    extract_paginated_items,
+    ok_envelope,
+    query_pairs_value,
+    upsert_query,
+    workflow_data_for_output_mode,
+    workflow_success_envelope,
+    workflow_success_envelope_with_data,
 };
+pub(in crate::api) use error::method_side_effecting;
+use input::{
+    parse_headers,
+    parse_key_values,
+    read_body,
+    split_path_and_inline_query,
+    write_json_output_file,
+    write_jsonl_items_output_file,
+};
+pub(in crate::api) use method::HttpMethod;
 use openapi::{
     api_coverage,
     command_next_actions,
@@ -88,7 +92,16 @@ use openapi::{
     synthetic_operation_id,
     workflow_alternatives,
 };
+use operations::WorkflowOperation;
 use render::print_output;
+use runtime_types::{
+    ApiRequestInput,
+    PreparedApiRequest,
+    RawPaginationOptions,
+    WorkflowCallResult,
+    WorkflowPaginationOptions,
+    WorkflowRequest,
+};
 use templates::{
     access_body_template,
     body_template,
@@ -101,12 +114,20 @@ use templates::{
     template_envelope,
     transfer_body_template,
 };
+use transport::{
+    read_api_response,
+    write_request_log,
+};
+pub(crate) use transport::{
+    request_id_from_headers,
+    response_body_value,
+};
+use workflow_options::ApiWorkflowOptions;
 use workflows::{
     access_request,
     account_request,
     assertions_next_actions,
     assertions_request,
-    compact_deployment_data,
     contracts_next_actions,
     contracts_request,
     deployments_request,
@@ -127,423 +148,6 @@ use workflows::{
     transfers_next_actions,
     transfers_request,
 };
-
-#[derive(Debug, thiserror::Error)]
-pub enum ApiCommandError {
-    #[error("Run `pcl auth login` first, or pass `--allow-unauthenticated`")]
-    NoAuthToken,
-
-    #[error(
-        "Stored auth token expired at {0}. Run `pcl auth refresh --toon` or `pcl auth login` again, or pass `--allow-unauthenticated` for public endpoints."
-    )]
-    ExpiredAuthToken(chrono::DateTime<chrono::Utc>),
-
-    #[error("Failed to refresh stored auth before retrying the API request: {0}")]
-    AuthRefresh(#[source] AuthError),
-
-    #[error("Invalid {kind} `{input}`. Expected KEY=VALUE.")]
-    InvalidKeyValue { kind: &'static str, input: String },
-
-    #[error("Invalid header name `{name}`: {source}")]
-    InvalidHeaderName {
-        name: String,
-        #[source]
-        source: reqwest::header::InvalidHeaderName,
-    },
-
-    #[error("Invalid header value for `{name}`: {source}")]
-    InvalidHeaderValue {
-        name: String,
-        #[source]
-        source: reqwest::header::InvalidHeaderValue,
-    },
-
-    #[error("Invalid API path `{0}`. Paths must start with `/`.")]
-    InvalidPath(String),
-
-    #[error("Failed to build API URL: {0}")]
-    Url(#[from] url::ParseError),
-
-    #[error("Failed to read body file `{path}`: {source}")]
-    BodyFile {
-        path: PathBuf,
-        #[source]
-        source: std::io::Error,
-    },
-
-    #[error("Failed to read request log `{path}`: {source}")]
-    RequestLog {
-        path: PathBuf,
-        #[source]
-        source: std::io::Error,
-    },
-
-    #[error("Failed to write output file `{path}`: {source}")]
-    OutputFile {
-        path: PathBuf,
-        #[source]
-        source: std::io::Error,
-    },
-
-    #[error("Failed to read request body from stdin: {0}")]
-    Stdin(std::io::Error),
-
-    #[error("Invalid JSON body: {0}")]
-    Json(#[from] serde_json::Error),
-
-    #[error("API request failed: {0}")]
-    Request(#[from] reqwest::Error),
-
-    #[error("API request failed with status {status} for {method} {path}")]
-    HttpStatus {
-        method: &'static str,
-        path: String,
-        status: u16,
-        request_id: Option<String>,
-        body: Box<Value>,
-    },
-
-    #[error("OpenAPI spec does not contain a paths object")]
-    MissingPaths,
-
-    #[error("No API operation matched `{0}`")]
-    OperationNotFound(String),
-
-    #[error("{message}")]
-    InvalidWorkflow { message: String },
-
-    #[error("{message}")]
-    InvalidWorkflowWithActions {
-        message: String,
-        next_actions: Vec<String>,
-    },
-}
-
-impl ApiCommandError {
-    pub fn code(&self) -> &'static str {
-        match self {
-            Self::NoAuthToken => "auth.no_token",
-            Self::ExpiredAuthToken(_) => "auth.expired_token",
-            Self::AuthRefresh(_) => "auth.refresh_failed",
-            Self::InvalidKeyValue { .. } => "input.invalid_key_value",
-            Self::InvalidHeaderName { .. } => "input.invalid_header_name",
-            Self::InvalidHeaderValue { .. } => "input.invalid_header_value",
-            Self::InvalidPath(_) => "input.invalid_path",
-            Self::Url(_) => "input.invalid_url",
-            Self::BodyFile { .. } => "input.body_file_read_failed",
-            Self::RequestLog { .. } => "request_log.read_failed",
-            Self::OutputFile { .. } => "output.file_write_failed",
-            Self::Stdin(_) => "input.stdin_read_failed",
-            Self::Json(_) => "input.invalid_json",
-            Self::Request(source) => {
-                match source.status().map(|status| status.as_u16()) {
-                    Some(400) => "api.bad_request",
-                    Some(401) => "auth.unauthorized",
-                    Some(403) => "auth.forbidden",
-                    Some(404) => "api.not_found",
-                    Some(422) => "api.validation_failed",
-                    Some(500..=599) => "api.server_error",
-                    _ => "network.request_failed",
-                }
-            }
-            Self::HttpStatus { status, .. } => {
-                match *status {
-                    400 => "api.bad_request",
-                    401 => "auth.unauthorized",
-                    403 => "auth.forbidden",
-                    404 => "api.not_found",
-                    422 => "api.validation_failed",
-                    500..=599 => "api.server_error",
-                    _ => "api.request_failed",
-                }
-            }
-            Self::MissingPaths => "openapi.missing_paths",
-            Self::OperationNotFound(_) => "openapi.operation_not_found",
-            Self::InvalidWorkflow { .. } | Self::InvalidWorkflowWithActions { .. } => {
-                "workflow.invalid_arguments"
-            }
-        }
-    }
-
-    pub fn recoverable(&self) -> bool {
-        !matches!(self, Self::MissingPaths)
-    }
-
-    pub fn next_actions(&self) -> Vec<String> {
-        match self {
-            Self::NoAuthToken | Self::ExpiredAuthToken(_) | Self::AuthRefresh(_) => {
-                vec![
-                    "pcl auth refresh --toon".to_string(),
-                    "pcl auth login".to_string(),
-                    "pcl api list --allow-unauthenticated --toon".to_string(),
-                ]
-            }
-            Self::InvalidPath(_) => {
-                vec![
-                    "pcl api list --toon".to_string(),
-                    "pcl api call get /views/public/incidents --allow-unauthenticated --toon"
-                        .to_string(),
-                ]
-            }
-            Self::InvalidKeyValue { kind, .. } => {
-                vec![format!(
-                    "Use --{kind} key=value, for example: pcl api call get /views/public/incidents --{kind} limit=5"
-                )]
-            }
-            Self::InvalidHeaderName { .. } | Self::InvalidHeaderValue { .. } => {
-                vec![
-                    "Use --header name=value, for example: --header x-cl-dev-mode=true".to_string(),
-                ]
-            }
-            Self::Json(_) => {
-                vec![
-                    "Use --field key=value for simple request bodies".to_string(),
-                    "Use --body-file request.json for nested request bodies".to_string(),
-                ]
-            }
-            Self::OperationNotFound(_) => {
-                vec![
-                    "pcl api list --toon".to_string(),
-                    "pcl api inspect get /views/public/incidents --toon".to_string(),
-                ]
-            }
-            Self::InvalidWorkflowWithActions { next_actions, .. } => next_actions.clone(),
-            Self::InvalidWorkflow { .. } => {
-                vec![
-                    "pcl projects mine".to_string(),
-                    "pcl schema list".to_string(),
-                    "pcl workflows".to_string(),
-                ]
-            }
-            Self::Request(source)
-                if matches!(
-                    source.status().map(|status| status.as_u16()),
-                    Some(401 | 403)
-                ) =>
-            {
-                vec![
-                    "pcl auth login".to_string(),
-                    "Use --allow-unauthenticated only for public endpoints".to_string(),
-                ]
-            }
-            Self::HttpStatus { status: 401, .. } => {
-                vec![
-                    "pcl auth refresh --toon".to_string(),
-                    "pcl auth login".to_string(),
-                    "Use --allow-unauthenticated only for public endpoints".to_string(),
-                ]
-            }
-            Self::HttpStatus { status: 403, .. } => {
-                vec![
-                    "Read error.http.body for the API-provided reason".to_string(),
-                    "Check whether the endpoint is enabled and your user has permission"
-                        .to_string(),
-                    "Use --allow-unauthenticated only for endpoints documented as public"
-                        .to_string(),
-                ]
-            }
-            Self::HttpStatus {
-                method,
-                path,
-                status: 400 | 422,
-                ..
-            } => {
-                vec![
-                    format!(
-                        "pcl api inspect {} {} --toon",
-                        method.to_ascii_lowercase(),
-                        path
-                    ),
-                    "pcl api manifest --toon".to_string(),
-                    "Read error.http.body for the rejected field details".to_string(),
-                ]
-            }
-            Self::HttpStatus { status: 404, .. } => {
-                vec![
-                    "Check the project ID, slug, or API path and retry".to_string(),
-                    "pcl projects mine".to_string(),
-                ]
-            }
-            Self::HttpStatus {
-                method,
-                status: status @ 500..=599,
-                request_id,
-                path,
-                ..
-            } => {
-                let mut actions = if mutation_outcome_ambiguous(method, *status) {
-                    vec![
-                        format!(
-                            "Do not retry immediately; inspect the target resource for {} {} to confirm whether the mutation applied",
-                            method.to_ascii_lowercase(),
-                            path
-                        ),
-                        "pcl requests list --toon".to_string(),
-                        "Read error.http.body for API-provided failure details".to_string(),
-                    ]
-                } else {
-                    vec![
-                        "Retry the same command once; server errors can be transient".to_string(),
-                        "pcl api manifest --toon".to_string(),
-                        "Read error.http.body for API-provided failure details".to_string(),
-                    ]
-                };
-                if let Some(request_id) = request_id {
-                    actions.push(format!(
-                        "Include request_id {request_id} when reporting this server error"
-                    ));
-                }
-                actions
-            }
-            Self::HttpStatus { .. } => {
-                vec![
-                    "pcl api manifest --toon".to_string(),
-                    "Read error.http.body for API-provided failure details".to_string(),
-                ]
-            }
-            Self::Request(source) if source.status().map(|status| status.as_u16()) == Some(404) => {
-                vec![
-                    "Check the project ID, slug, or API path and retry".to_string(),
-                    "pcl projects mine".to_string(),
-                ]
-            }
-            Self::Request(_) | Self::Url(_) => {
-                vec!["Check --api-url and your network connection, then retry".to_string()]
-            }
-            Self::BodyFile { .. } => {
-                vec!["Check --body-file path or pass --body directly".to_string()]
-            }
-            Self::RequestLog { .. } => {
-                vec![
-                    "pcl requests path --toon".to_string(),
-                    "Check request log permissions or move the PCL state directory".to_string(),
-                ]
-            }
-            Self::OutputFile { .. } => {
-                vec!["Check --output path permissions or choose a writable file".to_string()]
-            }
-            Self::Stdin(_) => vec!["Pipe a JSON body into --body-file -".to_string()],
-            Self::MissingPaths => {
-                vec!["Check that /api/v1/openapi returns an OpenAPI document".to_string()]
-            }
-        }
-    }
-
-    pub fn suggested_next_actions(&self) -> Vec<&'static str> {
-        match self {
-            Self::NoAuthToken | Self::ExpiredAuthToken(_) | Self::AuthRefresh(_) => {
-                vec!["refresh_or_login", "retry"]
-            }
-            Self::InvalidKeyValue { .. }
-            | Self::InvalidHeaderName { .. }
-            | Self::InvalidHeaderValue { .. }
-            | Self::InvalidPath(_)
-            | Self::Json(_)
-            | Self::InvalidWorkflow { .. }
-            | Self::InvalidWorkflowWithActions { .. } => vec!["fix_input", "retry"],
-            Self::OperationNotFound(_) | Self::MissingPaths => vec!["inspect_manifest"],
-            Self::Request(source) if source.status().map(|status| status.as_u16()) == Some(404) => {
-                vec!["check_ids", "retry"]
-            }
-            Self::Request(_) | Self::Url(_) => vec!["check_network", "retry"],
-            Self::BodyFile { .. } | Self::Stdin(_) => vec!["fix_body_input", "retry"],
-            Self::RequestLog { .. } => vec!["inspect_request_log", "retry"],
-            Self::OutputFile { .. } => vec!["fix_output_path", "retry"],
-            Self::HttpStatus { status: 401, .. } => vec!["refresh_or_login", "retry"],
-            Self::HttpStatus { status: 403, .. } => {
-                vec!["check_permissions", "inspect_response_body"]
-            }
-            Self::HttpStatus {
-                status: 400 | 422, ..
-            } => vec!["inspect_operation", "fix_request", "retry"],
-            Self::HttpStatus { status: 404, .. } => vec!["inspect_manifest", "check_ids"],
-            Self::HttpStatus { status: 429, .. } => vec!["retry_later", "reduce_request_rate"],
-            Self::HttpStatus {
-                method,
-                status: status @ 500..=599,
-                ..
-            } if mutation_outcome_ambiguous(method, *status) => {
-                vec!["reconcile_mutation", "contact_platform_with_request_id"]
-            }
-            Self::HttpStatus {
-                status: 500..=599, ..
-            } => {
-                vec![
-                    "retry_later",
-                    "export_project_incidents_with_errors",
-                    "contact_platform_with_request_id",
-                ]
-            }
-            Self::HttpStatus { .. } => vec!["inspect_response_body", "retry"],
-        }
-    }
-
-    pub fn json_envelope(&self) -> Value {
-        let mut error = Map::new();
-        error.insert("code".to_string(), json!(self.code()));
-        error.insert("message".to_string(), json!(self.to_string()));
-        error.insert("recoverable".to_string(), json!(self.recoverable()));
-
-        if let Self::HttpStatus {
-            method,
-            path,
-            status,
-            request_id,
-            body,
-        } = self
-        {
-            let outcome_ambiguous = mutation_outcome_ambiguous(method, *status);
-            if let Some(request_id) = request_id {
-                error.insert("request_id".to_string(), json!(request_id));
-            }
-            error.insert(
-                "http".to_string(),
-                json!({
-                    "method": method,
-                    "path": path,
-                    "status": status,
-                    "request_id": request_id,
-                    "body": body.as_ref(),
-                }),
-            );
-            error.insert(
-                "mutation".to_string(),
-                json!({
-                    "side_effecting": method_side_effecting(method),
-                    "outcome_ambiguous": outcome_ambiguous,
-                }),
-            );
-        }
-
-        let mut envelope = json!({
-            "status": "error",
-            "error": error,
-            "suggested_next_actions": self.suggested_next_actions(),
-            "next_actions": self.next_actions(),
-        });
-
-        if let Self::HttpStatus {
-            method,
-            path,
-            status,
-            request_id,
-            ..
-        } = self
-            && let Some(object) = envelope.as_object_mut()
-        {
-            object.insert("http_status".to_string(), json!(status));
-            object.insert("method".to_string(), json!(method));
-            object.insert("path".to_string(), json!(path));
-            object.insert("request_id".to_string(), json!(request_id));
-            object.insert(
-                "outcome_ambiguous".to_string(),
-                json!(mutation_outcome_ambiguous(method, *status)),
-            );
-        }
-
-        with_envelope_metadata(envelope)
-    }
-}
 
 #[derive(clap::Parser, Debug)]
 #[command(
@@ -570,61 +174,8 @@ pub struct ApiArgs {
     )]
     allow_unauthenticated: bool,
 
-    #[arg(
-        long = "dry-run",
-        global = true,
-        help = "Print the request plan without sending an API request"
-    )]
-    dry_run: bool,
-
     #[arg(skip = Cell::new(true))]
     refresh_after_401: Cell<bool>,
-}
-
-#[derive(clap::Args, Debug)]
-struct ApiWorkflowOptions {
-    #[arg(
-        long = "api-url",
-        env = "PCL_API_URL",
-        default_value = DEFAULT_PLATFORM_URL,
-        global = true,
-        help = "Base URL for the platform API"
-    )]
-    api_url: url::Url,
-
-    #[arg(
-        long,
-        global = true,
-        help = "Do not attach the stored bearer token to API requests"
-    )]
-    allow_unauthenticated: bool,
-
-    #[arg(
-        long = "dry-run",
-        global = true,
-        help = "Print the request plan without sending an API request"
-    )]
-    dry_run: bool,
-}
-
-impl ApiWorkflowOptions {
-    async fn run(
-        self,
-        command: ApiCommand,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        json_output: bool,
-    ) -> Result<(), ApiCommandError> {
-        ApiArgs {
-            command,
-            api_url: self.api_url,
-            allow_unauthenticated: self.allow_unauthenticated,
-            dry_run: self.dry_run,
-            refresh_after_401: Cell::new(true),
-        }
-        .run(config, cli_args, json_output)
-        .await
-    }
 }
 
 macro_rules! top_level_workflow_command {
@@ -858,143 +409,6 @@ enum ApiCommand {
         #[arg(long, help = "Write response body to a JSON file")]
         output: Option<PathBuf>,
     },
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq, ValueEnum)]
-enum HttpMethod {
-    Get,
-    Post,
-    Put,
-    Patch,
-    Delete,
-}
-
-impl HttpMethod {
-    fn as_str(self) -> &'static str {
-        match self {
-            Self::Get => "GET",
-            Self::Post => "POST",
-            Self::Put => "PUT",
-            Self::Patch => "PATCH",
-            Self::Delete => "DELETE",
-        }
-    }
-
-    fn openapi_key(self) -> &'static str {
-        match self {
-            Self::Get => "get",
-            Self::Post => "post",
-            Self::Put => "put",
-            Self::Patch => "patch",
-            Self::Delete => "delete",
-        }
-    }
-
-    fn reqwest(self) -> reqwest::Method {
-        match self {
-            Self::Get => reqwest::Method::GET,
-            Self::Post => reqwest::Method::POST,
-            Self::Put => reqwest::Method::PUT,
-            Self::Patch => reqwest::Method::PATCH,
-            Self::Delete => reqwest::Method::DELETE,
-        }
-    }
-}
-
-fn method_side_effecting(method: &str) -> bool {
-    !method.eq_ignore_ascii_case("GET") && !method.eq_ignore_ascii_case("HEAD")
-}
-
-fn mutation_outcome_ambiguous(method: &str, status: u16) -> bool {
-    method_side_effecting(method) && status >= 500
-}
-
-struct ApiRequestInput<'a> {
-    method: HttpMethod,
-    path: &'a str,
-    query: &'a [String],
-    header: &'a [String],
-    body: Option<&'a str>,
-    body_file: Option<&'a PathBuf>,
-    field: &'a [String],
-    require_auth: bool,
-}
-
-struct PreparedApiRequest<'a> {
-    attach_auth: bool,
-    method: HttpMethod,
-    url: &'a url::Url,
-    headers: &'a HeaderMap,
-    query: &'a [(String, String)],
-    body: Option<&'a Value>,
-}
-
-#[derive(Clone, Copy)]
-struct RawPaginationOptions<'a> {
-    item_field: &'a str,
-    start_page: u64,
-    limit: u64,
-    page_param: &'a str,
-    limit_param: &'a str,
-    max_pages: u64,
-}
-
-#[derive(Clone, Copy)]
-struct WorkflowPaginationOptions<'a> {
-    item_field: &'a str,
-    start_page: u64,
-    limit: u64,
-    max_pages: u64,
-}
-
-#[derive(Debug)]
-struct WorkflowCallResult {
-    body: Value,
-    request: Value,
-    response: Value,
-}
-
-#[derive(Clone, Debug)]
-struct WorkflowRequest {
-    method: HttpMethod,
-    path: String,
-    query: Vec<(String, String)>,
-    body: Option<String>,
-    require_auth: bool,
-    attach_auth: bool,
-    next_actions: Vec<String>,
-}
-
-impl WorkflowRequest {
-    fn get(
-        path: impl Into<String>,
-        require_auth: bool,
-        next_actions: impl IntoIterator<Item = impl Into<String>>,
-    ) -> Self {
-        Self::get_with_query(path, Vec::new(), require_auth, next_actions)
-    }
-
-    fn get_with_query(
-        path: impl Into<String>,
-        query: Vec<(String, String)>,
-        require_auth: bool,
-        next_actions: impl IntoIterator<Item = impl Into<String>>,
-    ) -> Self {
-        Self {
-            method: HttpMethod::Get,
-            path: path.into(),
-            query,
-            body: None,
-            require_auth,
-            attach_auth: require_auth,
-            next_actions: next_actions.into_iter().map(Into::into).collect(),
-        }
-    }
-
-    fn with_optional_auth(mut self) -> Self {
-        self.attach_auth = true;
-        self
-    }
 }
 
 #[derive(clap::Args, Debug)]
@@ -2412,1796 +1826,6 @@ top_level_workflow_command!(
     "Inspect project events and audit logs",
     "Examples:\n  pcl events --project <project-ref>\n  pcl events --project <project-ref> --audit-log\n\nCompatibility alias:\n  pcl api events ..."
 );
-
-impl ApiArgs {
-    pub async fn run(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        json_output: bool,
-    ) -> Result<(), ApiCommandError> {
-        let request_log_path = crate::request_log::request_log_path_for_args(cli_args);
-        match &self.command {
-            ApiCommand::Incidents(args) => {
-                let output = self
-                    .run_incidents(config, cli_args, args, &request_log_path)
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Projects(args) => {
-                let output = self
-                    .run_projects(config, cli_args, args, &request_log_path)
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Assertions(args) => {
-                let output = self
-                    .run_assertions(config, cli_args, args, &request_log_path)
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Search(args) => {
-                let output = self
-                    .run_search(config, cli_args, args, &request_log_path)
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Account(args) => {
-                if args.body_template {
-                    let output = template_envelope(body_template("empty_object"));
-                    print_output(&output, json_output)?;
-                    return Ok(());
-                }
-                let output = self
-                    .run_workflow(
-                        config,
-                        cli_args,
-                        "account",
-                        account_request(args)?,
-                        &request_log_path,
-                    )
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Contracts(args) => {
-                if args.body_template {
-                    let output = template_envelope(contracts_body_template(args));
-                    print_output(&output, json_output)?;
-                    return Ok(());
-                }
-                let output = self
-                    .run_contracts(config, cli_args, args, &request_log_path)
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Releases(args) => {
-                if args.body_template {
-                    let output = template_envelope(release_body_template(args));
-                    print_output(&output, json_output)?;
-                    return Ok(());
-                }
-                let output = self
-                    .run_releases(config, cli_args, args, &request_log_path)
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Deployments(args) => {
-                if args.body_template {
-                    let output = template_envelope(deployment_body_template(args));
-                    print_output(&output, json_output)?;
-                    return Ok(());
-                }
-                let output = self
-                    .run_deployments(config, cli_args, args, &request_log_path)
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Access(args) => {
-                if args.body_template {
-                    let output = template_envelope(access_body_template(args));
-                    print_output(&output, json_output)?;
-                    return Ok(());
-                }
-                let output = self
-                    .run_workflow(
-                        config,
-                        cli_args,
-                        "access",
-                        access_request(args)?,
-                        &request_log_path,
-                    )
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Integrations(args) => {
-                if args.body_template {
-                    let output = template_envelope(integration_body_template(args));
-                    print_output(&output, json_output)?;
-                    return Ok(());
-                }
-                let output = self
-                    .run_workflow(
-                        config,
-                        cli_args,
-                        "integrations",
-                        integrations_request(args)?,
-                        &request_log_path,
-                    )
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::ProtocolManager(args) => {
-                if args.body_template {
-                    let output = template_envelope(protocol_manager_body_template(args));
-                    print_output(&output, json_output)?;
-                    return Ok(());
-                }
-                let output = self
-                    .run_protocol_manager(config, cli_args, args, &request_log_path)
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Transfers(args) => {
-                if args.body_template {
-                    let output = template_envelope(transfer_body_template(args));
-                    print_output(&output, json_output)?;
-                    return Ok(());
-                }
-                let output = self
-                    .run_transfers(config, cli_args, args, &request_log_path)
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Events(args) => {
-                let output = self
-                    .run_workflow(
-                        config,
-                        cli_args,
-                        "events",
-                        events_request(args)?,
-                        &request_log_path,
-                    )
-                    .await?;
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Manifest => {
-                let output = ok_envelope(api_manifest());
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::List { filter, method } => {
-                let spec = self.fetch_openapi(config).await?;
-                let operations = list_operations(&spec, filter.as_deref(), *method)?;
-                let next_actions = next_actions_for_operations(&operations);
-                let output = json!({
-                    "status": "ok",
-                    "data": {
-                        "operations": operations,
-                    },
-                    "next_actions": next_actions,
-                });
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Inspect {
-                operation,
-                path,
-                full,
-            } => {
-                let spec = self.fetch_openapi(config).await?;
-                let inspected = inspect_operation(&spec, operation, path.as_deref(), *full)?;
-                let next_actions = command_next_actions(&inspected);
-                let output = json!({
-                    "status": "ok",
-                    "data": inspected,
-                    "next_actions": next_actions,
-                });
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Coverage { records, markdown } => {
-                let spec = self.fetch_openapi(config).await?;
-                let coverage =
-                    api_coverage(&spec, &request_log_path, *records, self.api_url.as_str())?;
-                if let Some(path) = markdown {
-                    write_api_coverage_markdown(path, &coverage)?;
-                }
-                let output = json!({
-                    "status": "ok",
-                    "data": coverage,
-                    "next_actions": [
-                        "pcl requests list --toon",
-                        "pcl api list --toon",
-                        "pcl api coverage --markdown api-coverage.md",
-                    ],
-                });
-                print_output(&output, json_output)?;
-            }
-            ApiCommand::Call {
-                method,
-                path,
-                query,
-                header,
-                body,
-                body_file,
-                field,
-                paginate,
-                all: _,
-                page,
-                limit,
-                page_param,
-                limit_param,
-                max_pages,
-                jsonl,
-                output,
-            } => {
-                if *jsonl && output.is_none() {
-                    return Err(ApiCommandError::InvalidWorkflow {
-                        message: "--jsonl requires --output".to_string(),
-                    });
-                }
-                let input = ApiRequestInput {
-                    method: *method,
-                    path,
-                    query,
-                    header,
-                    body: body.as_deref(),
-                    body_file: body_file.as_ref(),
-                    field,
-                    require_auth: self.raw_call_requires_auth(*method, path)?,
-                };
-                let pagination = paginate.as_ref().map(|item_field| {
-                    RawPaginationOptions {
-                        item_field,
-                        start_page: page.unwrap_or(1),
-                        limit: limit.unwrap_or(50),
-                        page_param: page_param.as_deref().unwrap_or("page"),
-                        limit_param: limit_param.as_deref().unwrap_or("limit"),
-                        max_pages: max_pages.unwrap_or(100),
-                    }
-                });
-                if self.dry_run {
-                    let output = dry_run_envelope(self.raw_call_plan(input, pagination, config)?);
-                    print_output(&output, json_output)?;
-                    return Ok(());
-                }
-                let (mut response, next_actions) = if let Some(pagination) = pagination {
-                    let response = self
-                        .call_api_paginated(config, cli_args, input, pagination, &request_log_path)
-                        .await?;
-                    (
-                        response,
-                        vec![
-                            "Adjust --limit or --max-pages if the result set was truncated"
-                                .to_string(),
-                            "Use --output results.json to save paginated data".to_string(),
-                            "pcl api manifest --toon".to_string(),
-                        ],
-                    )
-                } else {
-                    let response = self
-                        .call_api(config, cli_args, input, &request_log_path)
-                        .await?;
-                    (
-                        response,
-                        vec![
-                            "pcl api list --toon".to_string(),
-                            "pcl api manifest --toon".to_string(),
-                        ],
-                    )
-                };
-                if let Some(path) = output {
-                    if *jsonl {
-                        write_jsonl_items_output_file(path, &response)?;
-                    } else {
-                        let body = response.pointer("/response/body").unwrap_or(&response);
-                        write_json_output_file(path, body)?;
-                    }
-                    if let Some(object) = response.as_object_mut() {
-                        object.insert("output_path".to_string(), json!(path.display().to_string()));
-                    }
-                }
-                let output = json!({
-                    "status": "ok",
-                    "data": response,
-                    "next_actions": next_actions,
-                });
-                print_output(&output, json_output)?;
-            }
-        }
-
-        Ok(())
-    }
-
-    async fn call_api_paginated(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        input: ApiRequestInput<'_>,
-        pagination: RawPaginationOptions<'_>,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        if input.method.openapi_key() != "get" {
-            return Err(ApiCommandError::InvalidWorkflow {
-                message: "--paginate is only supported for GET requests".to_string(),
-            });
-        }
-        if input.body.is_some() || input.body_file.is_some() || !input.field.is_empty() {
-            return Err(ApiCommandError::InvalidWorkflow {
-                message: "--paginate cannot be used with request bodies".to_string(),
-            });
-        }
-        if pagination.limit == 0 {
-            return Err(ApiCommandError::InvalidWorkflow {
-                message: "--limit must be greater than zero".to_string(),
-            });
-        }
-        if pagination.max_pages == 0 {
-            return Err(ApiCommandError::InvalidWorkflow {
-                message: "--max-pages must be greater than zero".to_string(),
-            });
-        }
-
-        let (path, mut base_query) = split_path_and_inline_query(input.path)?;
-        base_query.extend(parse_key_values("query", input.query)?);
-        let url = self.api_url(&path)?;
-        let headers = parse_headers(input.header)?;
-        let operation_id = self.resolve_operation_id(config, input.method, &path).await;
-        self.ensure_request_auth(config, cli_args, input.require_auth)
-            .await?;
-        let client = self.http_client(
-            config,
-            input.require_auth && !self.allow_unauthenticated,
-            input.require_auth && !self.allow_unauthenticated,
-        )?;
-
-        let mut items = Vec::new();
-        let mut pages_fetched = 0_u64;
-        let mut last_page_count = 0_usize;
-
-        for offset in 0..pagination.max_pages {
-            let page = pagination.start_page + offset;
-            let mut page_query = base_query.clone();
-            upsert_query(&mut page_query, pagination.page_param, page.to_string());
-            upsert_query(
-                &mut page_query,
-                pagination.limit_param,
-                pagination.limit.to_string(),
-            );
-
-            let response = client
-                .get(url.clone())
-                .headers(headers.clone())
-                .query(&page_query)
-                .send()
-                .await?;
-            let status = response.status();
-            let request_id = request_id_from_headers(response.headers());
-            let content_type = response
-                .headers()
-                .get(reqwest::header::CONTENT_TYPE)
-                .and_then(|value| value.to_str().ok())
-                .unwrap_or_default()
-                .to_string();
-            let bytes = response.bytes().await?;
-            let body = response_body_value(&content_type, &bytes);
-            write_request_log(
-                request_log_path,
-                "raw_paginated",
-                input.method.as_str(),
-                &path,
-                status.as_u16(),
-                request_id.as_deref(),
-                operation_id.as_deref(),
-            );
-            if !status.is_success() {
-                return Err(ApiCommandError::HttpStatus {
-                    method: input.method.as_str(),
-                    path,
-                    status: status.as_u16(),
-                    request_id,
-                    body: Box::new(body),
-                });
-            }
-
-            let page_items =
-                extract_paginated_items(&body, pagination.item_field).ok_or_else(|| {
-                ApiCommandError::InvalidWorkflow {
-                    message: format!(
-                        "Could not find an array at `{}` or common pagination fields in response",
-                        pagination.item_field
-                    ),
-                }
-            })?;
-            last_page_count = page_items.len();
-            pages_fetched += 1;
-            items.extend(page_items);
-
-            if last_page_count < usize::try_from(pagination.limit).unwrap_or(usize::MAX) {
-                break;
-            }
-        }
-
-        let count = items.len();
-        Ok(json!({
-            "request": {
-                "method": input.method.as_str(),
-                "path": path,
-                "operation_id": operation_id,
-                "query": query_pairs_value(&base_query),
-                "pagination": {
-                    "field": pagination.item_field,
-                    "start_page": pagination.start_page,
-                    "limit": pagination.limit,
-                    "page_param": pagination.page_param,
-                    "limit_param": pagination.limit_param,
-                    "max_pages": pagination.max_pages,
-                }
-            },
-            "items": items,
-            "count": count,
-            "pages_fetched": pages_fetched,
-            "last_page_count": last_page_count,
-        }))
-    }
-
-    async fn run_incidents(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        args: &IncidentsArgs,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        let request = incidents_request(args)?;
-        if args.jsonl && args.output.is_none() {
-            return Err(ApiCommandError::InvalidWorkflow {
-                message: "--jsonl requires --output".to_string(),
-            });
-        }
-        if self.dry_run {
-            let pagination = args.all.then(|| {
-                json!({
-                    "enabled": true,
-                    "item_field": "incidents",
-                    "start_page": args.page.unwrap_or(1),
-                    "limit": args.limit.unwrap_or(50),
-                    "max_pages": args.max_pages.unwrap_or(100),
-                    "output": args.output.as_ref().map(|path| path.display().to_string()),
-                    "jsonl": args.jsonl,
-                })
-            });
-            return Ok(dry_run_envelope(
-                self.workflow_request_plan(&request, pagination, config),
-            ));
-        }
-        if args.all {
-            let mut data = self
-                .call_workflow_paginated(
-                    config,
-                    cli_args,
-                    request.clone(),
-                    WorkflowPaginationOptions {
-                        item_field: "incidents",
-                        start_page: args.page.unwrap_or(1),
-                        limit: args.limit.unwrap_or(50),
-                        max_pages: args.max_pages.unwrap_or(100),
-                    },
-                    request_log_path,
-                )
-                .await?;
-            if let Some(path) = &args.output {
-                if args.jsonl {
-                    write_jsonl_items_output_file(path, &data)?;
-                } else {
-                    write_json_output_file(path, &data)?;
-                }
-                if let Some(object) = data.as_object_mut() {
-                    object.insert("output_path".to_string(), json!(path.display().to_string()));
-                }
-            }
-            let mut next_actions = request.next_actions;
-            if args.output.is_none() {
-                next_actions.insert(
-                    0,
-                    "Use --output incidents.json to save large paginated results".to_string(),
-                );
-            }
-            return Ok(json!({
-                "status": "ok",
-                "data": data,
-                "next_actions": next_actions,
-            }));
-        }
-        let result = self
-            .call_workflow_result(config, cli_args, &request, request_log_path)
-            .await?;
-        let next_actions = incidents_next_actions(&result.body, args, request.next_actions);
-        Ok(workflow_success_envelope(result, next_actions))
-    }
-
-    async fn run_projects(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        args: &ProjectsArgs,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        if args.body_template {
-            return Ok(template_envelope(project_body_template(args)));
-        }
-        let request = projects_request(args)?;
-        self.run_prepared_workflow(
-            config,
-            cli_args,
-            "projects",
-            request,
-            request_log_path,
-            projects_next_actions,
-        )
-        .await
-    }
-
-    async fn run_assertions(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        args: &AssertionsArgs,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        if args.body_template {
-            return Ok(template_envelope(body_template("empty_object")));
-        }
-        let request = assertions_request(args)?;
-        self.run_prepared_workflow(
-            config,
-            cli_args,
-            "assertions",
-            request,
-            request_log_path,
-            |data, fallback| assertions_next_actions(data, args, fallback),
-        )
-        .await
-    }
-
-    async fn run_search(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        args: &SearchArgs,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        let request = search_request(args)?;
-        self.run_prepared_workflow(
-            config,
-            cli_args,
-            "search",
-            request,
-            request_log_path,
-            search_next_actions,
-        )
-        .await
-    }
-
-    async fn run_contracts(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        args: &ContractsArgs,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        let request = contracts_request(args)?;
-        self.run_prepared_workflow(
-            config,
-            cli_args,
-            "contracts",
-            request,
-            request_log_path,
-            |data, fallback| contracts_next_actions(data, args, fallback),
-        )
-        .await
-    }
-
-    async fn run_releases(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        args: &ReleasesArgs,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        let request = releases_request(args)?;
-        self.run_prepared_workflow(
-            config,
-            cli_args,
-            "releases",
-            request,
-            request_log_path,
-            |data, fallback| releases_next_actions(data, args, fallback),
-        )
-        .await
-    }
-
-    async fn run_deployments(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        args: &DeploymentsArgs,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        let request = deployments_request(args)?;
-        self.run_prepared_workflow(
-            config,
-            cli_args,
-            "deployments",
-            request,
-            request_log_path,
-            |_data, fallback| fallback,
-        )
-        .await
-    }
-
-    async fn run_transfers(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        args: &TransfersArgs,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        let request = transfers_request(args)?;
-        self.run_prepared_workflow(
-            config,
-            cli_args,
-            "transfers",
-            request,
-            request_log_path,
-            |data, fallback| transfers_next_actions(data, args, fallback),
-        )
-        .await
-    }
-
-    async fn run_protocol_manager(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        args: &ProtocolManagerArgs,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        let request = protocol_manager_request(args)?;
-        self.run_prepared_workflow(
-            config,
-            cli_args,
-            "protocol-manager",
-            request,
-            request_log_path,
-            |data, fallback| protocol_manager_next_actions(data, args, fallback),
-        )
-        .await
-    }
-
-    async fn run_workflow(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        workflow: &'static str,
-        request: WorkflowRequest,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        self.run_prepared_workflow(
-            config,
-            cli_args,
-            workflow,
-            request,
-            request_log_path,
-            |_data, fallback| fallback,
-        )
-        .await
-    }
-
-    async fn run_prepared_workflow<F>(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        workflow: &'static str,
-        request: WorkflowRequest,
-        request_log_path: &Path,
-        next_actions_for: F,
-    ) -> Result<Value, ApiCommandError>
-    where
-        F: FnOnce(&Value, Vec<String>) -> Vec<String>,
-    {
-        if self.dry_run {
-            return Ok(dry_run_envelope(
-                self.workflow_request_plan(&request, None, config),
-            ));
-        }
-        let result = self
-            .call_workflow_result(config, cli_args, &request, request_log_path)
-            .await?;
-        let next_actions = next_actions_for(&result.body, request.next_actions);
-        let data = workflow_data_for_output_mode(workflow, &result.body, cli_args.output_mode());
-        Ok(workflow_success_envelope_with_data(
-            result,
-            data,
-            next_actions,
-        ))
-    }
-
-    fn workflow_request_plan(
-        &self,
-        request: &WorkflowRequest,
-        pagination: Option<Value>,
-        config: &CliConfig,
-    ) -> Value {
-        let body = request.body.as_deref().map_or(Ok(Value::Null), |body| {
-            serde_json::from_str(body).map_err(ApiCommandError::Json)
-        });
-        let body = match body {
-            Ok(body) => body,
-            Err(error) => {
-                return json!({
-                    "dry_run": true,
-                    "valid": false,
-                    "error": {
-                        "code": error.code(),
-                        "message": error.to_string(),
-                    },
-                });
-            }
-        };
-
-        let destructive = request_is_destructive(request.method, &request.path);
-        json!({
-            "dry_run": true,
-            "valid": true,
-            "request": {
-                "method": request.method.as_str(),
-                "path": request.path.as_str(),
-                "query": query_pairs_value(&request.query),
-                "body": body,
-                "auth": self.auth_plan(request.require_auth, request.attach_auth, config),
-                "side_effecting": request.method != HttpMethod::Get,
-                "destructive": destructive,
-                "project_resolution": "not_performed",
-            },
-            "pagination": pagination,
-        })
-    }
-
-    fn raw_call_plan(
-        &self,
-        input: ApiRequestInput<'_>,
-        pagination: Option<RawPaginationOptions<'_>>,
-        config: &CliConfig,
-    ) -> Result<Value, ApiCommandError> {
-        let (path, mut query) = split_path_and_inline_query(input.path)?;
-        query.extend(parse_key_values("query", input.query)?);
-        let header = parse_key_values("header", input.header)?;
-        let body = request_body(input.body, input.body_file, input.field)?
-            .map(|body| serde_json::from_str::<Value>(&body))
-            .transpose()?;
-        let destructive = request_is_destructive(input.method, &path);
-
-        Ok(json!({
-            "dry_run": true,
-            "valid": true,
-            "request": {
-                "method": input.method.as_str(),
-                "path": path.as_str(),
-                "query": query_pairs_value(&query),
-                "headers": query_pairs_value(&header),
-                "body": body.unwrap_or(Value::Null),
-                "auth": self.auth_plan(input.require_auth, input.require_auth, config),
-                "side_effecting": input.method != HttpMethod::Get,
-                "destructive": destructive,
-            },
-            "pagination": pagination.map(|pagination| json!({
-                "enabled": true,
-                "item_field": pagination.item_field,
-                "start_page": pagination.start_page,
-                "limit": pagination.limit,
-                "page_param": pagination.page_param,
-                "limit_param": pagination.limit_param,
-                "max_pages": pagination.max_pages,
-            })),
-        }))
-    }
-
-    fn auth_plan(&self, require_auth: bool, attach_auth: bool, config: &CliConfig) -> Value {
-        let now = chrono::Utc::now();
-        let stored_token_present = config
-            .auth
-            .as_ref()
-            .is_some_and(|auth| !auth.access_token.trim().is_empty());
-        let stored_token_valid = config
-            .auth
-            .as_ref()
-            .is_some_and(|auth| !auth.access_token.trim().is_empty() && auth.expires_at > now);
-        let will_attach_stored_token =
-            attach_auth && !self.allow_unauthenticated && stored_token_valid;
-        json!({
-            "required": require_auth,
-            "will_attach_stored_token": will_attach_stored_token,
-            "stored_token_present": stored_token_present,
-            "stored_token_valid": stored_token_valid,
-            "allow_unauthenticated": self.allow_unauthenticated,
-        })
-    }
-
-    fn raw_call_requires_auth(
-        &self,
-        method: HttpMethod,
-        path: &str,
-    ) -> Result<bool, ApiCommandError> {
-        if self.allow_unauthenticated {
-            return Ok(false);
-        }
-        let (path, _) = split_path_and_inline_query(path)?;
-        Ok(!public_raw_call_path(method, &path))
-    }
-
-    async fn fetch_openapi(&self, config: &CliConfig) -> Result<Value, ApiCommandError> {
-        let url = self.api_url("/openapi")?;
-        let request = self.http_client(config, false, false)?.get(url);
-        let response = request.send().await?;
-        let status = response.status();
-        let request_id = request_id_from_headers(response.headers());
-        let content_type = response
-            .headers()
-            .get(reqwest::header::CONTENT_TYPE)
-            .and_then(|value| value.to_str().ok())
-            .unwrap_or_default()
-            .to_string();
-        let bytes = response.bytes().await?;
-        let body = response_body_value(&content_type, &bytes);
-        if !status.is_success() {
-            return Err(ApiCommandError::HttpStatus {
-                method: "GET",
-                path: "/openapi".to_string(),
-                status: status.as_u16(),
-                request_id,
-                body: Box::new(body),
-            });
-        }
-        Ok(body)
-    }
-
-    async fn try_refresh_after_401(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-    ) -> Result<bool, ApiCommandError> {
-        if !self.refresh_after_401.get() {
-            return Ok(false);
-        }
-
-        match refresh_stored_auth(config, &self.api_url, cli_args, true).await {
-            Ok(_) => Ok(true),
-            Err(AuthError::RefreshEndpointNotFound { .. }) => {
-                self.refresh_after_401.set(false);
-                Ok(false)
-            }
-            Err(error) => Err(ApiCommandError::AuthRefresh(error)),
-        }
-    }
-
-    async fn call_api(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        input: ApiRequestInput<'_>,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        let (path, mut query) = split_path_and_inline_query(input.path)?;
-        query.extend(parse_key_values("query", input.query)?);
-        let url = self.api_url(&path)?;
-        let headers = parse_headers(input.header)?;
-        let body = request_body(input.body, input.body_file, input.field)?;
-        let operation_id = self.resolve_operation_id(config, input.method, &path).await;
-        let requires_auth = input.require_auth && !self.allow_unauthenticated;
-        self.ensure_request_auth(config, cli_args, input.require_auth)
-            .await?;
-
-        let json_body = body
-            .as_deref()
-            .map(serde_json::from_str::<Value>)
-            .transpose()?;
-        let mut response = self
-            .send_api_request(
-                config,
-                PreparedApiRequest {
-                    attach_auth: requires_auth,
-                    method: input.method,
-                    url: &url,
-                    headers: &headers,
-                    query: &query,
-                    body: json_body.as_ref(),
-                },
-            )
-            .await?;
-        let mut status = response.status();
-        let request_id = request_id_from_headers(response.headers());
-        let response_headers = response
-            .headers()
-            .iter()
-            .filter_map(|(name, value)| {
-                value
-                    .to_str()
-                    .ok()
-                    .map(|value| (name.as_str().to_string(), json!(value)))
-            })
-            .collect::<serde_json::Map<_, _>>();
-        let content_type = response
-            .headers()
-            .get(reqwest::header::CONTENT_TYPE)
-            .and_then(|value| value.to_str().ok())
-            .unwrap_or_default()
-            .to_string();
-        let bytes = response.bytes().await?;
-        let mut body = response_body_value(&content_type, &bytes);
-        write_request_log(
-            request_log_path,
-            "raw",
-            input.method.as_str(),
-            &path,
-            status.as_u16(),
-            request_id.as_deref(),
-            operation_id.as_deref(),
-        );
-        if status.as_u16() == 401
-            && requires_auth
-            && self.try_refresh_after_401(config, cli_args).await?
-        {
-            response = self
-                .send_api_request(
-                    config,
-                    PreparedApiRequest {
-                        attach_auth: requires_auth,
-                        method: input.method,
-                        url: &url,
-                        headers: &headers,
-                        query: &query,
-                        body: json_body.as_ref(),
-                    },
-                )
-                .await?;
-            status = response.status();
-            let retry_request_id = request_id_from_headers(response.headers());
-            let retry_headers = response
-                .headers()
-                .iter()
-                .filter_map(|(name, value)| {
-                    value
-                        .to_str()
-                        .ok()
-                        .map(|value| (name.as_str().to_string(), json!(value)))
-                })
-                .collect::<serde_json::Map<_, _>>();
-            let retry_content_type = response
-                .headers()
-                .get(reqwest::header::CONTENT_TYPE)
-                .and_then(|value| value.to_str().ok())
-                .unwrap_or_default()
-                .to_string();
-            let retry_bytes = response.bytes().await?;
-            body = response_body_value(&retry_content_type, &retry_bytes);
-            write_request_log(
-                request_log_path,
-                "raw_retry_after_refresh",
-                input.method.as_str(),
-                &path,
-                status.as_u16(),
-                retry_request_id.as_deref(),
-                operation_id.as_deref(),
-            );
-            if !status.is_success() {
-                return Err(ApiCommandError::HttpStatus {
-                    method: input.method.as_str(),
-                    path,
-                    status: status.as_u16(),
-                    request_id: retry_request_id,
-                    body: Box::new(body),
-                });
-            }
-            return Ok(json!({
-                "request": {
-                    "method": input.method.as_str(),
-                    "path": path,
-                    "operation_id": operation_id,
-                    "query": query_pairs_value(&query),
-                    "retried_after_refresh": true,
-                },
-                "response": {
-                    "status": status.as_u16(),
-                    "success": status.is_success(),
-                    "request_id": retry_request_id,
-                    "headers": retry_headers,
-                    "body": body,
-                }
-            }));
-        }
-        if !status.is_success() {
-            return Err(ApiCommandError::HttpStatus {
-                method: input.method.as_str(),
-                path,
-                status: status.as_u16(),
-                request_id,
-                body: Box::new(body),
-            });
-        }
-
-        Ok(json!({
-            "request": {
-                "method": input.method.as_str(),
-                "path": path,
-                "operation_id": operation_id,
-                "query": query_pairs_value(&query),
-            },
-            "response": {
-                "status": status.as_u16(),
-                "success": status.is_success(),
-                "request_id": request_id,
-                "headers": response_headers,
-                "body": body,
-            }
-        }))
-    }
-
-    async fn call_workflow_result(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        request: &WorkflowRequest,
-        request_log_path: &Path,
-    ) -> Result<WorkflowCallResult, ApiCommandError> {
-        let requires_auth = request.require_auth && !self.allow_unauthenticated;
-        self.ensure_request_auth(config, cli_args, request.require_auth)
-            .await?;
-        let attach_auth = self.workflow_attach_auth(request, config);
-        let path = self
-            .normalize_project_path(
-                config,
-                &request.path,
-                attach_auth,
-                requires_auth,
-                request_log_path,
-            )
-            .await?;
-        let url = self.api_url(&path)?;
-        let json_body = if let Some(body) = &request.body {
-            Some(
-                self.normalize_request_body(
-                    config,
-                    &path,
-                    body,
-                    attach_auth,
-                    requires_auth,
-                    request_log_path,
-                )
-                .await?,
-            )
-        } else {
-            None
-        };
-        let mut response = self
-            .send_workflow_request(config, request, &url, json_body.as_ref())
-            .await?;
-        let mut status = response.status();
-        let mut request_id = request_id_from_headers(response.headers());
-        let content_type = response
-            .headers()
-            .get(reqwest::header::CONTENT_TYPE)
-            .and_then(|value| value.to_str().ok())
-            .unwrap_or_default()
-            .to_string();
-        let bytes = response.bytes().await?;
-        let mut body = response_body_value(&content_type, &bytes);
-        write_request_log(
-            request_log_path,
-            "workflow",
-            request.method.as_str(),
-            &path,
-            status.as_u16(),
-            request_id.as_deref(),
-            None,
-        );
-        let mut retried_after_refresh = false;
-        if status.as_u16() == 401
-            && requires_auth
-            && self.try_refresh_after_401(config, cli_args).await?
-        {
-            response = self
-                .send_workflow_request(config, request, &url, json_body.as_ref())
-                .await?;
-            status = response.status();
-            request_id = request_id_from_headers(response.headers());
-            let retry_content_type = response
-                .headers()
-                .get(reqwest::header::CONTENT_TYPE)
-                .and_then(|value| value.to_str().ok())
-                .unwrap_or_default()
-                .to_string();
-            let retry_bytes = response.bytes().await?;
-            body = response_body_value(&retry_content_type, &retry_bytes);
-            retried_after_refresh = true;
-            write_request_log(
-                request_log_path,
-                "workflow_retry_after_refresh",
-                request.method.as_str(),
-                &path,
-                status.as_u16(),
-                request_id.as_deref(),
-                None,
-            );
-        }
-        if !status.is_success() {
-            return Err(ApiCommandError::HttpStatus {
-                method: request.method.as_str(),
-                path,
-                status: status.as_u16(),
-                request_id,
-                body: Box::new(body),
-            });
-        }
-        Ok(WorkflowCallResult {
-            body,
-            request: json!({
-                "method": request.method.as_str(),
-                "path": path,
-                "query": query_pairs_value(&request.query),
-                "auth": self.auth_plan(request.require_auth, request.attach_auth, config),
-                "side_effecting": request.method != HttpMethod::Get,
-                "destructive": request_is_destructive(request.method, &request.path),
-                "retried_after_refresh": retried_after_refresh,
-            }),
-            response: json!({
-                "status": status.as_u16(),
-                "success": true,
-                "request_id": request_id,
-                "fetched_at": chrono::Utc::now().to_rfc3339(),
-            }),
-        })
-    }
-
-    async fn call_workflow_paginated(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        request: WorkflowRequest,
-        pagination: WorkflowPaginationOptions<'_>,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        if request.method.openapi_key() != "get" {
-            return Err(ApiCommandError::InvalidWorkflow {
-                message: "--all is only supported for GET list workflows".to_string(),
-            });
-        }
-        if pagination.limit == 0 {
-            return Err(ApiCommandError::InvalidWorkflow {
-                message: "--limit must be greater than zero".to_string(),
-            });
-        }
-        if pagination.max_pages == 0 {
-            return Err(ApiCommandError::InvalidWorkflow {
-                message: "--max-pages must be greater than zero".to_string(),
-            });
-        }
-
-        let mut items = Vec::new();
-        let mut pages_fetched = 0_u64;
-        let mut last_page_count = 0_usize;
-
-        for offset in 0..pagination.max_pages {
-            let page = pagination.start_page + offset;
-            let mut page_request = request.clone();
-            upsert_query(&mut page_request.query, "page", page.to_string());
-            upsert_query(
-                &mut page_request.query,
-                "limit",
-                pagination.limit.to_string(),
-            );
-            let data = self
-                .call_workflow_result(config, cli_args, &page_request, request_log_path)
-                .await?
-                .body;
-            let page_items =
-                extract_paginated_items(&data, pagination.item_field).ok_or_else(|| {
-                ApiCommandError::InvalidWorkflow {
-                    message: format!(
-                        "Could not find an array at `{}` or common pagination fields in response",
-                        pagination.item_field
-                    ),
-                }
-            })?;
-            last_page_count = page_items.len();
-            pages_fetched += 1;
-            items.extend(page_items);
-
-            if last_page_count < usize::try_from(pagination.limit).unwrap_or(usize::MAX) {
-                break;
-            }
-        }
-
-        let count = items.len();
-        Ok(json!({
-            "items": items,
-            "count": count,
-            "pages_fetched": pages_fetched,
-            "start_page": pagination.start_page,
-            "limit": pagination.limit,
-            "max_pages": pagination.max_pages,
-            "last_page_count": last_page_count,
-        }))
-    }
-
-    async fn normalize_request_body(
-        &self,
-        config: &CliConfig,
-        path: &str,
-        body: &str,
-        attach_auth: bool,
-        require_auth: bool,
-        request_log_path: &Path,
-    ) -> Result<Value, ApiCommandError> {
-        let mut json_body: Value = serde_json::from_str(body)?;
-        if path == "/projects/saved"
-            && let Some(project_ref) = json_body.get("project_id").and_then(Value::as_str)
-            && project_ref.parse::<uuid::Uuid>().is_err()
-        {
-            let project_id = self
-                .resolve_project_id(
-                    config,
-                    project_ref,
-                    attach_auth,
-                    require_auth,
-                    request_log_path,
-                )
-                .await?;
-            if let Some(object) = json_body.as_object_mut() {
-                object.insert("project_id".to_string(), Value::String(project_id));
-            }
-        }
-        Ok(json_body)
-    }
-
-    async fn normalize_project_path(
-        &self,
-        config: &CliConfig,
-        path: &str,
-        attach_auth: bool,
-        require_auth: bool,
-        request_log_path: &Path,
-    ) -> Result<String, ApiCommandError> {
-        let Some((prefix, project_ref, suffix)) = project_segment(path) else {
-            return Ok(path.to_string());
-        };
-        if project_ref.parse::<uuid::Uuid>().is_ok() {
-            return Ok(path.to_string());
-        }
-        let project_id = self
-            .resolve_project_id(
-                config,
-                project_ref,
-                attach_auth,
-                require_auth,
-                request_log_path,
-            )
-            .await?;
-        Ok(format!("{prefix}{project_id}{suffix}"))
-    }
-
-    async fn resolve_project_id(
-        &self,
-        config: &CliConfig,
-        project_ref: &str,
-        attach_auth: bool,
-        require_auth: bool,
-        request_log_path: &Path,
-    ) -> Result<String, ApiCommandError> {
-        let path = format!("/projects/resolve/{project_ref}");
-        let url = self.api_url(&path)?;
-        let client = self.http_client(config, attach_auth, require_auth)?;
-        let response = client.get(url).send().await?;
-        let status = response.status();
-        let request_id = request_id_from_headers(response.headers());
-        let content_type = response
-            .headers()
-            .get(reqwest::header::CONTENT_TYPE)
-            .and_then(|value| value.to_str().ok())
-            .unwrap_or_default()
-            .to_string();
-        let bytes = response.bytes().await?;
-        let response = response_body_value(&content_type, &bytes);
-        write_request_log(
-            request_log_path,
-            "workflow_project_resolution",
-            "GET",
-            &path,
-            status.as_u16(),
-            request_id.as_deref(),
-            Some("get_projects_resolve_project_ref"),
-        );
-        if !status.is_success() {
-            return Err(ApiCommandError::HttpStatus {
-                method: "GET",
-                path,
-                status: status.as_u16(),
-                request_id,
-                body: Box::new(response),
-            });
-        }
-        response
-            .get("project_id")
-            .or_else(|| response.get("projectId"))
-            .or_else(|| response.get("id"))
-            .and_then(Value::as_str)
-            .map(ToString::to_string)
-            .ok_or_else(|| {
-                ApiCommandError::InvalidWorkflow {
-                    message: format!("Could not resolve project reference `{project_ref}`"),
-                }
-            })
-    }
-
-    async fn ensure_request_auth(
-        &self,
-        config: &mut CliConfig,
-        cli_args: &CliArgs,
-        require_auth: bool,
-    ) -> Result<(), ApiCommandError> {
-        if self.allow_unauthenticated || !require_auth {
-            return Ok(());
-        }
-        let Some(auth) = &config.auth else {
-            return Err(ApiCommandError::NoAuthToken);
-        };
-        let now = chrono::Utc::now();
-        let seconds_remaining = (auth.expires_at - now).num_seconds();
-        if auth.expires_at <= now || seconds_remaining <= crate::config::AUTH_EXPIRES_SOON_SECONDS {
-            refresh_stored_auth(config, &self.api_url, cli_args, false)
-                .await
-                .map_err(ApiCommandError::AuthRefresh)?;
-        }
-        Ok(())
-    }
-
-    async fn send_api_request(
-        &self,
-        config: &CliConfig,
-        request: PreparedApiRequest<'_>,
-    ) -> Result<reqwest::Response, ApiCommandError> {
-        let client = self.http_client(config, request.attach_auth, request.attach_auth)?;
-        let mut builder = client
-            .request(request.method.reqwest(), request.url.clone())
-            .headers(request.headers.clone());
-        if !request.query.is_empty() {
-            builder = builder.query(request.query);
-        }
-        if let Some(body) = request.body {
-            builder = builder.json(body);
-        }
-        Ok(builder.send().await?)
-    }
-
-    async fn send_workflow_request(
-        &self,
-        config: &CliConfig,
-        request: &WorkflowRequest,
-        url: &url::Url,
-        body: Option<&Value>,
-    ) -> Result<reqwest::Response, ApiCommandError> {
-        let requires_auth = request.require_auth && !self.allow_unauthenticated;
-        let attach_auth = self.workflow_attach_auth(request, config);
-        let client = self.http_client(config, attach_auth, requires_auth)?;
-        let mut builder = client.request(request.method.reqwest(), url.clone());
-        if !request.query.is_empty() {
-            builder = builder.query(&request.query);
-        }
-        if let Some(body) = body {
-            builder = builder.json(body);
-        }
-        Ok(builder.send().await?)
-    }
-
-    fn workflow_attach_auth(&self, request: &WorkflowRequest, config: &CliConfig) -> bool {
-        if self.allow_unauthenticated {
-            return false;
-        }
-        if request.require_auth {
-            return true;
-        }
-        request.attach_auth
-            && config.auth.as_ref().is_some_and(|auth| {
-                !auth.access_token.trim().is_empty() && auth.expires_at > chrono::Utc::now()
-            })
-    }
-
-    fn http_client(
-        &self,
-        config: &CliConfig,
-        attach_auth: bool,
-        require_auth: bool,
-    ) -> Result<reqwest::Client, ApiCommandError> {
-        let mut headers = HeaderMap::new();
-        headers.insert(
-            HeaderName::from_static("api-version"),
-            HeaderValue::from_static("1"),
-        );
-
-        if attach_auth && let Some(auth) = &config.auth {
-            if auth.expires_at <= chrono::Utc::now() {
-                return Err(ApiCommandError::ExpiredAuthToken(auth.expires_at));
-            }
-
-            let value = format!("Bearer {}", auth.access_token);
-            let value = HeaderValue::from_str(&value).map_err(|source| {
-                ApiCommandError::InvalidHeaderValue {
-                    name: "authorization".to_string(),
-                    source,
-                }
-            })?;
-            headers.insert(reqwest::header::AUTHORIZATION, value);
-        } else if require_auth {
-            return Err(ApiCommandError::NoAuthToken);
-        }
-
-        reqwest::Client::builder()
-            .default_headers(headers)
-            .build()
-            .map_err(ApiCommandError::Request)
-    }
-
-    async fn resolve_operation_id(
-        &self,
-        config: &CliConfig,
-        method: HttpMethod,
-        path: &str,
-    ) -> Option<String> {
-        let spec = self.fetch_openapi(config).await.ok()?;
-        let operations = list_operations(&spec, None, Some(method)).ok()?;
-        operations
-            .into_iter()
-            .find(|operation| openapi_path_matches(&operation.path, path))
-            .map(|operation| operation.operation_id)
-    }
-
-    fn api_url(&self, path: &str) -> Result<url::Url, ApiCommandError> {
-        if !path.starts_with('/') {
-            return Err(ApiCommandError::InvalidPath(path.to_string()));
-        }
-
-        let mut url = self.api_url.clone();
-        url.set_path(&format!("/api/v1{path}"));
-        Ok(url)
-    }
-}
-
-fn split_path_and_inline_query(
-    input: &str,
-) -> Result<(String, Vec<(String, String)>), ApiCommandError> {
-    if !input.starts_with('/') {
-        return Err(ApiCommandError::InvalidPath(input.to_string()));
-    }
-    let Some((path, query)) = input.split_once('?') else {
-        return Ok((input.to_string(), Vec::new()));
-    };
-    if path.is_empty() || !path.starts_with('/') {
-        return Err(ApiCommandError::InvalidPath(input.to_string()));
-    }
-    let query = url::form_urlencoded::parse(query.as_bytes())
-        .map(|(key, value)| (key.into_owned(), value.into_owned()))
-        .collect();
-    Ok((path.to_string(), query))
-}
-
-pub(crate) fn request_id_from_headers(headers: &HeaderMap) -> Option<String> {
-    [
-        "x-request-id",
-        "x-correlation-id",
-        "x-amzn-requestid",
-        "cf-ray",
-        "request-id",
-    ]
-    .into_iter()
-    .find_map(|name| {
-        headers
-            .get(name)
-            .and_then(|value| value.to_str().ok())
-            .filter(|value| !value.is_empty())
-            .map(ToString::to_string)
-    })
-}
-
-fn write_request_log(
-    request_log_path: &Path,
-    kind: &str,
-    method: &str,
-    path: &str,
-    status: u16,
-    request_id: Option<&str>,
-    operation_id: Option<&str>,
-) {
-    #[cfg(not(test))]
-    {
-        let _ = crate::request_log::append_request_record_at(
-            request_log_path,
-            &json!({
-                "timestamp": chrono::Utc::now().to_rfc3339(),
-                "kind": kind,
-                "method": method,
-                "path": path,
-                "status": status,
-                "success": (200..=299).contains(&status),
-                "request_id": request_id,
-                "operation_id": operation_id,
-            }),
-        );
-    }
-    #[cfg(test)]
-    let _ = (
-        request_log_path,
-        kind,
-        method,
-        path,
-        status,
-        request_id,
-        operation_id,
-    );
-}
-
-pub(crate) fn response_body_value(content_type: &str, bytes: &[u8]) -> Value {
-    if content_type.contains("application/json") {
-        return serde_json::from_slice(bytes).unwrap_or_else(|_| {
-            json!({
-                "parse_error": "response declared JSON but could not be parsed",
-                "raw": String::from_utf8_lossy(bytes),
-            })
-        });
-    }
-
-    serde_json::from_slice(bytes)
-        .unwrap_or_else(|_| json!(String::from_utf8_lossy(bytes).to_string()))
-}
-
-fn ok_envelope(data: Value) -> Value {
-    with_envelope_metadata(json!({
-        "status": "ok",
-        "data": data,
-        "next_actions": [
-            "pcl api list",
-            "pcl api inspect get_views_public_incidents",
-            "pcl api call get /views/public/incidents --query limit=5 --allow-unauthenticated",
-        ],
-    }))
-}
-
-fn dry_run_envelope(data: Value) -> Value {
-    let auth_required = data
-        .pointer("/request/auth/required")
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-    let allow_unauthenticated = data
-        .pointer("/request/auth/allow_unauthenticated")
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-    let stored_token_valid = data
-        .pointer("/request/auth/stored_token_valid")
-        .and_then(Value::as_bool)
-        .unwrap_or(false);
-    let method = data
-        .pointer("/request/method")
-        .and_then(Value::as_str)
-        .unwrap_or_default();
-    let mut next_actions = if auth_required && !allow_unauthenticated && !stored_token_valid {
-        vec![
-            "pcl auth ensure --toon",
-            "Authenticate before removing --dry-run",
-        ]
-    } else {
-        vec![
-            "Remove --dry-run to execute this request",
-            "Use --toon for agent consumption or --json for strict JSON parsing",
-        ]
-    };
-    if method_side_effecting(method) {
-        next_actions.push("Use --body-template when constructing mutation bodies");
-    }
-    with_envelope_metadata(json!({
-        "status": "ok",
-        "data": data,
-        "next_actions": next_actions,
-    }))
-}
-
-fn workflow_success_envelope(result: WorkflowCallResult, next_actions: Vec<String>) -> Value {
-    with_envelope_metadata(json!({
-        "status": "ok",
-        "data": result.body,
-        "request": result.request,
-        "response": result.response,
-        "next_actions": next_actions,
-    }))
-}
-
-fn workflow_success_envelope_with_data(
-    result: WorkflowCallResult,
-    data: Value,
-    next_actions: Vec<String>,
-) -> Value {
-    with_envelope_metadata(json!({
-        "status": "ok",
-        "data": data,
-        "request": result.request,
-        "response": result.response,
-        "next_actions": next_actions,
-    }))
-}
-
-fn workflow_data_for_output_mode(workflow: &str, data: &Value, output_mode: OutputMode) -> Value {
-    match (workflow_output_policy(workflow), output_mode) {
-        (WorkflowOutputPolicy::MachineRawHumanCompactArtifacts, OutputMode::Human) => {
-            compact_deployment_data(data)
-        }
-        _ => data.clone(),
-    }
-}
-
-fn request_is_destructive(method: HttpMethod, path: &str) -> bool {
-    method == HttpMethod::Delete
-        || path.contains("/delete")
-        || path.contains("/remove")
-        || path.contains("/reject")
-        || path.contains("/logout")
-}
-
-fn query_pairs_value(query: &[(String, String)]) -> Value {
-    Value::Array(
-        query
-            .iter()
-            .map(|(name, value)| json!({ "name": name, "value": value }))
-            .collect(),
-    )
-}
-
-fn upsert_query(query: &mut Vec<(String, String)>, name: &str, value: String) {
-    if let Some((_, existing)) = query.iter_mut().find(|(key, _)| key == name) {
-        *existing = value;
-    } else {
-        query.push((name.to_string(), value));
-    }
-}
-
-fn extract_paginated_items(value: &Value, preferred_field: &str) -> Option<Vec<Value>> {
-    if let Some(items) = array_at_path(value, preferred_field) {
-        return Some(items.to_vec());
-    }
-    for path in [
-        "items",
-        "incidents",
-        "results",
-        "data.items",
-        "data.incidents",
-        "data.results",
-        "data",
-    ] {
-        if let Some(items) = array_at_path(value, path) {
-            return Some(items.to_vec());
-        }
-    }
-    value.as_array().cloned()
-}
-
-fn array_at_path<'a>(value: &'a Value, path: &str) -> Option<&'a [Value]> {
-    let mut current = value;
-    for segment in path.split('.') {
-        if segment.is_empty() {
-            continue;
-        }
-        current = current.get(segment)?;
-    }
-    current.as_array().map(Vec::as_slice)
-}
-
-fn parse_key_values(
-    kind: &'static str,
-    entries: &[String],
-) -> Result<Vec<(String, String)>, ApiCommandError> {
-    entries
-        .iter()
-        .map(|entry| {
-            let (key, value) = entry.split_once('=').ok_or_else(|| {
-                ApiCommandError::InvalidKeyValue {
-                    kind,
-                    input: entry.clone(),
-                }
-            })?;
-            Ok((key.to_string(), value.to_string()))
-        })
-        .collect()
-}
-
-fn parse_headers(entries: &[String]) -> Result<HeaderMap, ApiCommandError> {
-    let mut headers = HeaderMap::new();
-
-    for entry in entries {
-        let (name, value) = entry.split_once('=').ok_or_else(|| {
-            ApiCommandError::InvalidKeyValue {
-                kind: "header",
-                input: entry.clone(),
-            }
-        })?;
-        let header_name = HeaderName::from_str(name).map_err(|source| {
-            ApiCommandError::InvalidHeaderName {
-                name: name.to_string(),
-                source,
-            }
-        })?;
-        let header_value = HeaderValue::from_str(value).map_err(|source| {
-            ApiCommandError::InvalidHeaderValue {
-                name: name.to_string(),
-                source,
-            }
-        })?;
-        headers.insert(header_name, header_value);
-    }
-
-    Ok(headers)
-}
-
-fn read_body(
-    body: Option<&str>,
-    body_file: Option<&PathBuf>,
-) -> Result<Option<String>, ApiCommandError> {
-    if let Some(body) = body {
-        return Ok(Some(body.to_string()));
-    }
-
-    if let Some(path) = body_file {
-        if path.as_os_str() == "-" {
-            let mut body = String::new();
-            std::io::stdin()
-                .read_to_string(&mut body)
-                .map_err(ApiCommandError::Stdin)?;
-            return Ok(Some(body));
-        }
-
-        return fs::read_to_string(path).map(Some).map_err(|source| {
-            ApiCommandError::BodyFile {
-                path: path.clone(),
-                source,
-            }
-        });
-    }
-
-    Ok(None)
-}
-
-fn write_json_output_file(path: &PathBuf, value: &Value) -> Result<(), ApiCommandError> {
-    let body = serde_json::to_string_pretty(value)?;
-    fs::write(path, body).map_err(|source| {
-        ApiCommandError::OutputFile {
-            path: path.clone(),
-            source,
-        }
-    })
-}
-
-fn write_jsonl_items_output_file(path: &PathBuf, value: &Value) -> Result<(), ApiCommandError> {
-    let items = value
-        .get("items")
-        .and_then(Value::as_array)
-        .ok_or_else(|| {
-            ApiCommandError::InvalidWorkflow {
-                message: "--jsonl output requires paginated data with an items array".to_string(),
-            }
-        })?;
-    let mut body = String::new();
-    for item in items {
-        body.push_str(&serde_json::to_string(item)?);
-        body.push('\n');
-    }
-    fs::write(path, body).map_err(|source| {
-        ApiCommandError::OutputFile {
-            path: path.clone(),
-            source,
-        }
-    })
-}
 
 #[cfg(test)]
 mod tests;
