+++ crates/pcl/core/src/api/workflows/incidents.rs
@@ -3,16 +3,13 @@ use super::{
         ApiCommandError,
         HttpMethod,
         IncidentsArgs,
+        WorkflowOperation,
         WorkflowRequest,
-        definitions::{
-            WorkflowActionDefinition,
-            WorkflowDefinition,
-            WorkflowOutputPolicy,
-        },
     },
     first_string_field,
     push_query,
     required_arg,
+    workflow_operation_with_body,
 };
 use serde_json::Value;
 
@@ -47,46 +44,65 @@ pub(in crate::api) fn incidents_request(
     if let Some(incident_id) = &args.incident_id {
         if args.retry_trace {
             let tx_id = required_arg(args.tx_id.as_deref(), "--tx-id")?;
-            return Ok(WorkflowRequest {
-                method: HttpMethod::Post,
-                path: format!("/incidents/{incident_id}/transactions/{tx_id}/trace/retry"),
-                query,
-                body: Some("{}".to_string()),
-                require_auth: true,
-                attach_auth: true,
-                next_actions: vec![format!(
+            return workflow_operation_with_body(
+                WorkflowOperation::new(
+                    HttpMethod::Post,
+                    "post_incidents_incident_id_transactions_tx_id_trace_retry",
+                )
+                .path_param("incident_id", incident_id)
+                .path_param("tx_id", &tx_id),
+                true,
+                Some("{}".to_string()),
+                vec![format!(
                     "pcl incidents --incident-id {incident_id} --tx-id {tx_id}"
                 )],
-            });
+            );
         }
-        let path = if let Some(tx_id) = &args.tx_id {
-            format!("/views/incidents/{incident_id}/transactions/{tx_id}/trace")
+        let request = if let Some(tx_id) = &args.tx_id {
+            WorkflowRequest::from_operation(
+                WorkflowOperation::new(
+                    HttpMethod::Get,
+                    "get_views_incidents_incident_id_transactions_tx_id_trace",
+                )
+                .path_param("incidentId", incident_id)
+                .path_param("txId", tx_id),
+                query,
+                None,
+                true,
+                vec![
+                    "pcl incidents --limit 5".to_string(),
+                    "pcl api inspect get_views_incidents_incident_id_transactions_tx_id_trace"
+                        .to_string(),
+                ],
+            )?
         } else {
-            format!("/views/incidents/{incident_id}")
+            WorkflowRequest::from_operation(
+                WorkflowOperation::new(HttpMethod::Get, "get_views_incidents_incident_id")
+                    .path_param("incidentId", incident_id),
+                query,
+                None,
+                true,
+                vec![
+                    "pcl incidents --limit 5".to_string(),
+                    "pcl api inspect get_views_incidents_incident_id".to_string(),
+                ],
+            )?
         };
-        let next_actions = vec![
-            "pcl incidents --limit 5".to_string(),
-            format!("pcl api inspect get {}", path),
-        ];
-        return Ok(WorkflowRequest::get_with_query(
-            path,
-            query,
-            true,
-            next_actions,
-        ));
+        return Ok(request);
     }
 
     if let Some(project_id) = &args.project_id {
         if args.stats {
-            let path = format!("/projects/{project_id}/incidents/stats");
-            return Ok(WorkflowRequest::get_with_query(
-                path,
+            return WorkflowRequest::from_operation(
+                WorkflowOperation::new(HttpMethod::Get, "get_projects_project_id_incidents_stats")
+                    .path_param("project_id", project_id),
                 query,
+                None,
                 true,
                 vec![format!(
                     "pcl incidents --project-id {project_id} --limit 10"
                 )],
-            ));
+            );
         }
         push_query(&mut query, "assertionId", args.assertion_id.as_deref());
         push_query(
@@ -97,30 +113,32 @@ pub(in crate::api) fn incidents_request(
         push_query(&mut query, "environment", args.environment.as_deref());
         push_query(&mut query, "fromDate", args.from_date.as_deref());
         push_query(&mut query, "toDate", args.to_date.as_deref());
-        let path = format!("/views/projects/{project_id}/incidents");
-        return Ok(WorkflowRequest::get_with_query(
-            path,
+        return WorkflowRequest::from_operation(
+            WorkflowOperation::new(HttpMethod::Get, "get_views_projects_project_id_incidents")
+                .path_param("projectId", project_id),
             query,
+            None,
             true,
             vec![
                 format!("pcl assertions --project-id {project_id}"),
                 "pcl incidents --limit 5".to_string(),
             ],
-        ));
+        );
     }
 
     push_query(&mut query, "network", args.network);
     push_query(&mut query, "sort", args.sort.as_deref());
     push_query(&mut query, "devMode", args.dev_mode.as_deref());
-    Ok(WorkflowRequest::get_with_query(
-        "/views/public/incidents",
+    WorkflowRequest::from_operation(
+        WorkflowOperation::new(HttpMethod::Get, "get_views_public_incidents"),
         query,
+        None,
         false,
         vec![
             "pcl incidents --project-id <project-id> --limit 10".to_string(),
             "pcl projects list --limit 10".to_string(),
         ],
-    ))
+    )
 }
 
 pub(in crate::api) fn incidents_next_actions(
@@ -136,7 +154,15 @@ pub(in crate::api) fn incidents_next_actions(
                 .and_then(Value::as_array)
                 .and_then(|transactions| transactions.first())
                 .and_then(|transaction| {
-                    first_string_field(transaction, &["transaction_hash", "id", "tx_id"])
+                    first_string_field(
+                        transaction,
+                        &[
+                            "id",
+                            "tx_id",
+                            "invalidating_transaction_id",
+                            "invalidatingTransactionId",
+                        ],
+                    )
                 })
         {
             return vec![
@@ -154,14 +180,16 @@ pub(in crate::api) fn incidents_next_actions(
     })
 }
 
-pub(in crate::api) const DEFINITION: WorkflowDefinition = WorkflowDefinition {
-    name: "incidents",
+workflow_definition!(
+    "incidents",
     command: "pcl incidents [--project-id <id>] [--incident-id <id>] [--stats] [--limit <n>] [--all --output <file>]",
     description: "List public incidents, project incidents, fetch all incident pages, inspect incident detail, incident stats, or incident trace.",
     output: "incident data from /views/public/incidents, /views/projects/{projectId}/incidents, /views/incidents/{incidentId}, or /projects/{project_id}/incidents/stats",
-    output_policy: WorkflowOutputPolicy::MachineRaw,
-    legacy_examples: &[],
-    actions: &[
+    policy: MachineRaw,
+    legacy_examples: [
+
+    ],
+    actions: [
         action!("list_public", false, "GET", "/views/public/incidents", "pcl incidents --limit 5", optional: ["--page", "--limit", "--network", "--sort", "--dev-mode", "--all", "--max-pages", "--output"]),
         action!("list_project", true, "GET", "/views/projects/{projectId}/incidents", "pcl incidents --project <project-ref> --all --limit 50 --output incidents.json", required: ["--project"], optional: ["--page", "--limit", "--assertion-id", "--adopter-id", "--environment", "--from", "--to", "--all", "--max-pages", "--output"]),
         action!("stats", true, "GET", "/projects/{project_id}/incidents/stats", "pcl incidents --project <project-ref> --stats", required: ["--project"]),
@@ -169,4 +197,4 @@ pub(in crate::api) const DEFINITION: WorkflowDefinition = WorkflowDefinition {
         action!("trace", true, "GET", "/views/incidents/{incidentId}/transactions/{txId}/trace", "pcl incidents --incident-id <incident-id> --tx-id <invalidating-transaction-id>", required: ["--incident-id", "--tx-id"]),
         action!("retry_trace", true, "POST", "/incidents/{incident_id}/transactions/{tx_id}/trace/retry", "pcl incidents --incident-id <incident-id> --tx-id <tx-id> --retry-trace", required: ["--incident-id", "--tx-id"], body_template: "empty_object"),
     ],
-};
+);
